Re-organize the analysis code as well as the resulting SMT program.
Both now look more like a regular interpretation of the underlying P4
program, just done symbolically.

The analysis code maintains no mutable state, except for the scope of an
action body.

State is organized as helper structs. Analysis code is organized as
functions, each matching a unique construct from our IR. These functions
return a symbolic expression, representing the symbolic interpretation
of the corresponding IR construct, or an error message if invalid
constructs are encountered.

The SMT program consists of a single constraint per table. The
constraint is a conjunction of alias definitions (one per entry), and a
single if-elseif-else construct, matching on every entry and ending with
a special 'dropped' alias. The body of every match is a symbolic
expression constraining the parameters of the corresponding action and
asserting that action's body symbolically.